+++ code/system.ui/react.ui.media/src/types.mts
@@ -1,4 +1,5 @@
 export type * from './ui/Audio.Player/types.mjs';
+export type * from './ui/Audio.Waveform/types.mjs';
 export type * from './ui/Concept.Player/types.mjs';
 export type * from './ui/MediaStream/types';
 export type * from './ui/RecordButton/types.mjs';
