+++ code/system.ui/react.ui.media/src/ui/Audio.Waveform/types.mts
@@ -1,0 +1,1 @@
+export {};
